requirements.txt
agate==1.9.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.6.2.post1
apache-airflow==2.10.3
apache-airflow-providers-common-compat==1.2.1
apache-airflow-providers-common-io==1.4.2
apache-airflow-providers-common-sql==1.19.0
apache-airflow-providers-fab==1.5.0
apache-airflow-providers-ftp==3.11.1
apache-airflow-providers-http==4.13.2
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-smtp==1.8.0
apache-airflow-providers-sqlite==3.9.0
apispec==6.7.1
argcomplete==3.5.1
asgiref==3.8.1
asn1crypto==1.5.1
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
backports.tarfile==1.2.0
blinker==1.9.0
cachelib==0.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
coloredlogs==14.0
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.2.0
cron-descriptor==1.4.5
croniter==3.0.4
cryptography==43.0.3
cycler==0.12.1
daff==1.3.46
dagster==1.9.3
dagster-dbt==0.25.3
dagster-graphql==1.9.3
dagster-pipes==1.9.3
dagster-webserver==1.9.3
dbt-adapters==1.7.0
dbt-common==1.10.0
dbt-core==1.8.8
dbt-extractor==0.5.1
dbt-postgres==1.8.2
dbt-semantic-interfaces==0.5.1
dbt-snowflake==1.8.4
deepdiff==7.0.1
Deprecated==1.2.15
dill==0.3.9
dnspython==2.7.0
docstring_parser==0.16
duckdb==1.1.3
email_validator==2.2.0
fastjsonschema==2.21.0
filelock==3.16.1
Flask==2.2.5
Flask-AppBuilder==4.5.2
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
fonttools==4.50.0
frozenlist==1.5.0
fsspec==2024.10.0
gdown==4.4.0
google-re2==1.1.20240702
googleapis-common-protos==1.66.0
gql==3.5.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.68.0
grpcio-health-checking==1.62.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
humanfriendly==10.0
idna==3.10
imgaug==0.4.0
importlib-metadata==6.11.0
importlib_resources==6.4.5
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupytext==1.16.4
keyring==25.5.0
kiwisolver==1.4.5
lazy-object-proxy==1.10.0
leather==0.4.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
Logbook==1.5.3
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mashumaro==3.14
matplotlib==3.8.3
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
minimal-snowplow-tracker==0.0.2
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.1.0
nbformat==5.10.4
networkx==3.4.2
numpy==2.1.3
opencv-python==4.9.0.80
opentelemetry-api==1.28.1
opentelemetry-exporter-otlp==1.28.1
opentelemetry-exporter-otlp-proto-common==1.28.1
opentelemetry-exporter-otlp-proto-grpc==1.28.1
opentelemetry-exporter-otlp-proto-http==1.28.1
opentelemetry-proto==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
ordered-set==4.1.0
orjson==3.10.12
packaging==24.2
pandas==2.2.3
parsedatetime==2.6
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.3.6
pluggy==1.5.0
prison==0.2.1
propcache==0.2.0
protobuf==4.25.5
psutil==6.1.0
psycopg2-binary==2.9.10
pyarrow==18.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.10.0
pyOpenSSL==24.2.1
pyreadline3==3.5.4
python-daemon==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-nvd3==0.16.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.6.0
rpds-py==0.21.0
setproctitle==1.3.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowflake-connector-python==3.12.3
sortedcontainers==2.4.0
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlglot==25.32.0
sqlglotrs==0.3.0
sqlparse==0.5.2
starlette==0.41.3
structlog==24.4.0
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
text-unidecode==1.3
time-machine==2.16.0
tomli==2.1.0
tomlkit==0.13.2
toposort==1.10
tqdm==4.67.1
traitlets==5.14.3
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
universal_pathlib==0.2.5
urllib3==2.2.3
uvicorn==0.32.1
watchdog==5.0.3
watchfiles==1.0.0
websockets==14.1
Werkzeug==2.2.3
wirerope==0.4.8
wrapt==1.16.0
WTForms==3.2.1
yarl==1.17.2
zipp==3.21.0
